# backend/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: kubeport API
  version: "0.1.0"
  license:
    name: MIT
    identifier: MIT
  description: |
    kubeport turns Kubernetes resources into a self-service catalog: admins
    publish versioned templates, everyone else deploys them through generated
    forms.

    **This API is a UX layer, not an authorization boundary.** Every write that
    touches a cluster is performed with the caller's own OIDC id_token, forwarded
    unchanged to that cluster's apiserver. Kubernetes RBAC decides what actually
    happens, so a 201 from kubeport only means kubeport accepted the request —
    the cluster may still refuse it.

    ## Two ways in

    The Go API described here listens on `:8080` **inside the cluster only**
    (`ClusterIP`; the Ingress routes all external traffic to the frontend). It
    takes an OIDC id_token as a bearer token.

    Browsers instead talk to the Next.js BFF at `https://<host>/api/v1/*`, which
    authenticates from an httpOnly session cookie and injects the bearer token
    server-side. It **ignores any `Authorization` header you send**. Paths and
    bodies are otherwise identical to this spec.

    So a script port-forwards to the backend Service and sends a bearer
    id_token. Driving the browser session is **not** a client path: the BFF
    ignores `Authorization`, and the session token never leaves the server (it
    is held encrypted in the `sessions` table, not in the cookie).

...
    which, and issue #34 tracks the gap.

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/clusters:
    get:
      tags: [clusters]
      operationId: listClusters
      summary: List clusters available as deploy targets
      description: |
        Returns only what is needed to target a deploy. `api_url`, `ca_bundle`
        and `oidc_issuer_url` are deliberately **not** included for any caller,
        admin included — see issue #52. Register a cluster to see them once, in
        the POST response.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [clusters]
                properties:
                  clusters:
                    type: array
                    items: { $ref: "#/components/schemas/ClusterSummary" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [clusters]
      operationId: createCluster
      summary: Register a cluster as a deploy target
      description: |
        Admin only, and refused for demo-domain callers. There is no
        cluster-registration screen in the admin UI, so this endpoint is the
        supported path — see `deploy/helm/kubeport/README.md`, "After install".
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateClusterRequest" }
      responses:
        "201":
          description: |
            Created. This is the only response that carries the connection
            fields back, so keep it if you want to verify what was stored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cluster" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: A cluster with that name already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /v1/releases/{id}/logs:
    get:
      tags: [releases]
      operationId: streamReleaseLogs
      summary: Follow pod logs (Server-Sent Events)
      description: |
        Long-lived `text/event-stream`. Event types:

        | event | data |
        |---|---|
        | `log`   | `{"time": <unix ms>, "pod": "<name>", "text": "<line>"}`. `time` is when the container wrote the line, from the kubelet's own stamp — not when the server forwarded it (#131); replayed history carries past times, and under `instance=all` frame order and `time` order can differ, so do not read `time` as progress. A stamped line also carries an SSE `id:`, the resume point below. **On a named-instance stream** it is the line's instant, always UTC with exactly nine fractional digits (`2026-09-09T07:36:36.120000000Z`), so ids sort as strings in time order. **On an `instance=all` stream** it is a per-pod cursor: `pod@instant` for every pod the stream has a position for, sorted by pod and comma-joined, each instant in that same form (`web-1@2026-09-09T07:36:36.120000000Z,web-2@2026-09-09T07:35:00.900000000Z`). Hand a cursor back whole, the last one received; do not compare cursors as strings |
        | `ping`  | unix seconds, every 15s to keep the connection open |
        | `error` | a `Problem`, same schema and same `ErrorKind` vocabulary as any other error. `title` is `rbac-denied` (403, the caller may read the release but not its pods' logs — permanent), `cluster-auth-denied` (502, the cluster refused the forwarded token — permanent), or `k8s-error` (502, transport or anything else — retryable) |
        | `end`   | `{"reason": "all pods stopped emitting"}` — a fixed string. The stream is over; the server closes immediately after. Always the last frame a client receives |
        | `replay` | `{}`. The request carried a resume point (`Last-Event-ID` or `since`) that this stream does not apply — an unreadable id, or a cursor on an `instance=all` stream past its pod bound — so the logs follow from the beginning. Discard what was received before it. Sent first, before any `log` frame, and only then (#172). Its SSE `id:` is `-`, which replaces the resume point the client held; sent back (as `Last-Event-ID` or `since`) it starts over again rather than resuming |

        Authentication failures before the stream opens are ordinary JSON
        Problems; once it is open, failures arrive as `error` events — and
        since #82 the two are the same shape, so one parser covers both. A
        cluster call that fails while listing the release's pods — before the
        handshake — carries the same three kinds as the table above rather than
        collapsing into `k8s-error`: the handshake is not supposed to change
        what a failure is called, and a client that gates a retry on the kind
        was being told a permanent refusal was worth retrying (#134). The
        frame carries `request_id`, which is the only way to reach the reason:
        `detail` is deliberately generic because client-go's text names the
        apiserver's address, the namespace and the pod, and this frame is
        rendered verbatim in the log pane (#108).

        An `error` frame does not by itself mean the stream is over — with
        `instance=all` the handler keeps following the healthy pods after
        writing one. The stream ends when every pod has stopped emitting, and
        the `end` frame is what says so. Errors buffered at that moment are
        flushed first — they used to race the close and were dropped about half
        the time.

        **Treat `end` as the signal to stop, not the connection closing.**
        Nothing else distinguishes a finished stream from a dropped one:
        WHATWG gives `EventSource` no way to tell them apart, so a client that
        waits for the close instead reconnects on the browser's own 3s timer,
        indefinitely (#162) — and on an `instance=all` stream, which has no
        resume point (below), is served the whole container log every time.
        Close the connection yourself when `end` arrives.

        A connection that closes **without** `end` was dropped, not finished,
        and reopening is the right response. The only paths that omit the frame
        are the ones where the client has already gone, so its absence never
        means "finished, quietly".

        **Reopening resumes rather than replays.** Each stamped `log` frame
        carries an SSE `id:`, so a standard `EventSource` sends the last one back
        as `Last-Event-ID` on its automatic reconnect and picks up where it left
        off: no client code, and no duplicates in a pane the browser does not
        clear (#107). A caller with no `EventSource` keeping track for it passes
        the same value as `?since=`, which wins if both are present.

        **A named instance resumes from one instant.** The name is in the
        request, so it means the same pod on every reconnect, and if that pod is
        gone the answer is `404 no-pods` rather than a different pod inheriting
        its position.

        **`instance=all` resumes each pod from its own position** (#172). One
        instant cannot stand for progress through a set: pods are followed
        concurrently and merged without ordering, so the newest line belongs to
        whichever wrote last, and resuming every pod from it would skip —
        silently and for good — whatever a slower one had not reached. So the id
        is a cursor holding every pod's position, and on reconnect each pod picks
        up from its own. The set can change between connections: a pod the
        cursor does not name — one that started since — is sent from the
        beginning, and a position for a pod that is gone is dropped rather than
        handed to the pod that replaced it. For the same reason a plain instant
        in `?since=` on an `all` stream is `400 validation-error`, and a plain
        instant in `Last-Event-ID` there (left over from a single-instance stream
        on the same `EventSource`) is ignored.

        The cursor rides on every frame, so it is bounded: past 8 pods an `all`
        stream ignores a cursor and replays from the beginning as before,
        announcing it with a `replay` frame. Its stamped frames then carry the
        id `-`, which is not a cursor, so a later reconnect — even after the
        release is back under the bound — starts over too rather than resuming
        from a cursor older than what was replayed. Resending lines is the
        lesser cost.

        A line the kubelet could not stamp carries no id either. Its `time` is
        the server's clock so the row has something to render, and letting that
        into the cursor would claim the client had read up to *now*.

        The cluster's own filter is whole seconds, so it returns everything
        from the second you asked about; the server trims that overlap against
        the exact instant before sending, so a resume is duplicate-free without
        the caller deduplicating. A line the kubelet could not stamp is always
        sent — a possible duplicate beats a line that silently disappears.

        Sending no resume point gets the container log from the beginning,
        which is what a first open wants and what the Reconnect button in the
        web UI does, since it empties the pane at the same time.

        `reason` is human-readable text, not a branch key — it carries no closed
        vocabulary and may be reworded, and like an `error` frame's `detail` it
        must never carry the cluster's own words (#108). When a stream ended
        because of a failure, the thing to branch on is the `title` of the
        preceding `error` frame. If `reason` ever needs to vary by meaning, give
        it an enum beside `ErrorKind` with the same build-time guard rather than
        changing this prose.

        `end` is additive: a client that registers no listener for it never sees
        the frame — `onmessage` receives only unnamed events — and behaves
        exactly as it did before, so nothing that worked breaks. Parsers that
        render every `data:` line verbatim will print the payload, so branch on
        the event name rather than on the payload.

        This endpoint has no way to tell a client that a *pod* is done while
        others continue; `end` is about the whole stream.

        **A stream is also ended by the server after a fixed lifetime** — one
        hour by default, which an install can change — and that ending sends
        **no** `end` frame, because it is not a finish (#169). Nothing else
        would ever close a stream the reader keeps open, and the stream does
        not re-check its caller after the handshake, so without this a revoked
        permission or an expired session went on receiving logs for as long as
        the tab stayed open. Treat it like any drop: reconnect. A browser does
        so on its own, through a refreshed token and a new authorization; a
        named instance resumes from its instant and `instance=all` from its
        per-pod cursor (replaying only past the cursor's pod bound).

        A caller may also hold only so many streams open at once (16 by
        default, which an install can change); past that, opening another is
        `429 too-many-streams` — see the `RateLimited` response.
      parameters:
        - $ref: "#/components/parameters/ReleaseID"
        - name: instance
          in: query
          description: Pod name, or `all` (default) to follow every pod.
          schema: { type: string, default: all }
        - name: since
          in: query
          required: false
          description: |
            Resume point, normally the `id:` of the last `log` frame received.
            Lines at or before it are not sent. Omit to get the container log
            from the beginning.

            Its shape follows `instance`. On a named instance it is an RFC3339
            instant. On `instance=all` — the default — it is the per-pod cursor
            that stream's frames carry: `pod@instant` pairs joined by commas
            (at most 8), each pod resuming from its own instant (#172). A pod
            the cursor does not name is sent from the beginning.

            A browser does not need this — its `EventSource` sends
            `Last-Event-ID` by itself. This parameter is for callers that do
            not, and it takes precedence over that header when both are given.

            `-` — the id of a `replay` frame, and of an `instance=all` stream
            past its pod bound — is accepted on either view and starts the
            stream over, with a `replay` frame.

            A plain instant with `instance=all` is `400 validation-error`
            rather than silently applied to every pod: that is the resume that
            skips whatever a slower pod had not reached. (A plain instant in
            `Last-Event-ID` on an `all` stream is ignored, not refused — the
            browser keeps it across an instance switch.)

            An unparseable `since` is `400 validation-error`: the caller wrote
            it, and a 200 with the whole log would let them keep sending it.
            An unparseable `Last-Event-ID` is **not** an error — it starts the
            stream from the beginning instead, after a `replay` frame. No
            caller typed that value; the
            browser re-attaches it on every automatic reconnect and nothing on
            the page can clear it, so refusing it would leave the tab with a log
            pane that fails on every reconnect until reloaded.
          # Not `format: date-time`: on `instance=all` it is a per-pod cursor.
          schema: { type: string }
      responses:
        "200":
          description: Stream opened.
          content:
            text/event-stream:
              schema: { type: string }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: |
            No such release, or no pod matched `instance` — `title` is
            `not-found` or `no-pods`. `no-pods` is an observation, not a
            verdict: a pod that is still Pending or pulling, and a Job between
            runs, both land here, so it is worth retrying.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An OIDC **id_token** (not an access token) from an issuer this install
        trusts. The backend checks signature, issuer and `aud`, then forwards
        the same token to the target cluster.

        How to obtain one non-interactively is not solved for production IdPs —
        see `docs/machine-clients.md` and issue #34.
  parameters:
...
    ReleaseID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  headers:
    RequestId:
      description: |
        Echoed on **every** response, and equal to `Problem.request_id` on an
        error. An inbound `X-Request-Id` matching `^[A-Za-z0-9._-]{1,64}$` is
        adopted, so a caller's own id survives the BFF hop; anything else is
        replaced with a generated one.
      schema:
        type: string
        pattern: "^[A-Za-z0-9._-]{1,64}$"
  responses:
    Unauthenticated:
      description: |
        Missing, malformed or rejected token. Through the BFF this also covers a
        missing or expired session cookie.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://kubeport.io/errors/unauthenticated
            title: unauthenticated
            status: 401
            detail: missing bearer token
            request_id: 4d3f0a2c-7c1e-4a6b-9f10-2f5f1f0b8e21
    Forbidden:
      description: |
        `title` says who refused: `rbac-denied` is kubeport's own check or the
        cluster's RBAC; `demo-restricted` is the demo gate — a management write
        demo accounts never get, template creation or a version's publish,
        deprecate or undeprecate on an
        install without demo.allowTemplateCreate, or a read or change of a release that no demo
        account owns. A template on the other side of the demo line is not a
        403: reading, changing or deploying it answers 404 like a missing one
        (#238, #244).
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ValidationError:
      description: The request body or a parameter did not validate.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    PayloadTooLarge:
      description: |
        The request body is over the 4 MiB cap (`payload-too-large`, #128).
        Until #128 this was a `400 validation-error` whose `detail` was Go's
        `http: request body too large`; it has its own kind because the
        response is the opposite one — a validation error is fixed by
        correcting the body, this one only by sending less of it. **Do not
        retry the same request.** Shrink it, or split a template into several.

        A `Content-Length` over the cap is refused before routing and
        authentication, on every operation and method — so an unauthenticated
        caller sees 413 rather than 401, and a GET that declares such a length
        is refused even though no GET reads a body. A body of unknown length
        (chunked) is cut off at the cap while the handler reads it, with the
        same response. Only operations that take a request body list this
        response, since those are the ones a well-formed client can reach it on.

        The server does not drain the rest of an oversized body and closes the
        connection after answering, so a client still uploading may see the
        connection drop before it reads this. Measure before sending.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://kubeport.io/errors/payload-too-large
            title: payload-too-large
            status: 413
            detail: request body exceeds 4 MiB
            request_id: 4d3f0a2c-7c1e-4a6b-9f10-2f5f1f0b8e21
    RateLimited:
      description: |
        Over the per-caller budget (`rate-limited`). The budget is a token
        bucket keyed by OIDC subject, in-process and per replica — it exists to
        bound how often a caller can make kubeport talk to a cluster's control
        plane, not to price the API (issue #73).

        `Retry-After` says how many seconds to wait; honour it rather than
        retrying at once or sleeping an arbitrary constant. The demo's two Dex
        accounts are static, so concurrent demo visitors share one bucket.

        Budgets are per route group, not one per caller: reading a release
        (`GET /v1/releases/{id}`, 240/min) does not spend the budget shared by
        the SSAR, cluster-openapi and log-stream routes (60/min); creating,
        updating and deleting a release share one of their own (30/min), so
        switching verbs does not reset it; and the template authoring and
        preview routes have their own. The route decides
        the budget; `X-RateLimit-Limit` repeats its size, and two budgets are
        both 60/min. The `X-RateLimit-*` pair is sent only on a 429 — a
        successful response carries no remaining count, so a polling client
        paces itself. On the demo, every visitor of one Dex account spends the
        same 240/min of release reads.

        `too-many-streams` is the other 429, and only
        `GET /v1/releases/{id}/logs` gives it: the caller already holds as many
        log streams **open** as it may (#169). The bucket above prices opening a
        stream and nothing after, so it cannot bound how many a caller keeps —
        and each keeps a goroutine per pod and an apiserver connection alive.
        Branch on `title`: this one clears when one of the caller's streams
        closes, not after a set time, so close a stream you no longer need
        rather than only backing off. `Retry-After` is still present, as a
        backoff hint rather than a prediction, and the `X-RateLimit-*` pair is
        **not** — it describes the per-minute budget and would be wrong about a
        count of open streams. The cap is per caller, so the demo's shared
        accounts share it too — and each sign-in to a demo account also has a
        cap of 8 of its own (#200), so one sign-in cannot hold the whole pool.
        Signing in again brings a new cap, so this raises what holding the pool
        costs rather than preventing it.
      headers:
        Retry-After:
          description: Seconds to wait before retrying. Always present on a 429.
          schema: { type: integer, minimum: 1 }
        X-RateLimit-Limit:
          description: Requests allowed per minute for this caller.
          schema: { type: integer }
        X-RateLimit-Remaining:
          description: Always `0` here — the header is emitted only on refusal.
          schema: { type: integer }
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    UpstreamError:
      description: |
        The call to the target cluster failed or was refused (`k8s-error`). An
        RBAC denial from the apiserver arrives here rather than as a 403, so
        this is not automatically a transient condition — check `detail` before
        retrying.

        `detail` is the apiserver's own words **only when the apiserver
        answered** — a Forbidden, Invalid, NotFound, AlreadyExists, Conflict or
        Unauthorized verdict. `User "x" cannot create deployments` is why the
        deploy failed and belongs on screen. When the call never got there, the
        transport error names the cluster's internal address, so it goes to the
        log and `detail` is only the operation that failed (issue #49).
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    InternalError:
      description: |
        Something failed on kubeport's side. `title` is `internal` for a
        database, rendering or k8s-client-construction failure, or
        `cluster-config` when the stored registration's TLS material is not
        usable.

        **`detail` no longer carries the underlying error** (issue #49). A pgx
        connect failure spelled out `host=… user=… database=…`, and a k8s
        transport failure carried the apiserver's address — the same value
        `GET /v1/clusters` deliberately withholds (#52). The reason is in the
        server log; quote `request_id` to have it looked up.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    ClusterSummary:
      type: object
      description: What GET /v1/clusters returns. Connection fields are omitted.
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        display_name: { type: [string, "null"] }
        default_namespace: { type: [string, "null"] }
    Cluster:
      allOf:
        - $ref: "#/components/schemas/ClusterSummary"
        - type: object
          description: Full record, returned only by POST /v1/clusters.
          properties:
            api_url: { type: string, format: uri }
            ca_bundle:
              type: [string, "null"]
              description: |
                PEM text, not base64. Null only on a cluster registered before
                validation existed, or with `KBP_DEV_ALLOW_INSECURE_CLUSTERS`.
            oidc_issuer_url: { type: string, format: uri }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    CreateClusterRequest:
      type: object
      # ca_bundle is listed as required because the endpoint refuses without
      # it. The one exception is a backend running with
      # KBP_DEV_ALLOW_INSECURE_CLUSTERS=true, which is local dev only.
      required: [name, api_url, ca_bundle, oidc_issuer_url]
      properties:
        name:
          type: string
          minLength: 1
          description: Slug releases refer to. Must be unique.
        display_name: { type: string }
        api_url:
          type: string
          format: uri
          description: For an in-cluster target, `https://kubernetes.default.svc`.
        ca_bundle:
          type: string
          description: |
            PEM text, not base64. **Required**, and validated here: an empty or
            unparseable bundle is a 400 (issue #96).

            It used to be optional, and omitting it made the backend connect
            with TLS verification disabled while still returning 201 — on the
            very path that forwards the user's id_token to the apiserver. It now
            fails at registration rather than at the first deploy, where the
            reason reached only the pod log and the UI reported "cluster
            unreachable", pointing the operator at the network.

            `KBP_DEV_ALLOW_INSECURE_CLUSTERS=true` on the backend restores the
            old behaviour for a local kind cluster. Local dev only.
        oidc_issuer_url: { type: string, format: uri }
        default_namespace:
          type: string
          description: Pre-selected in the deploy form.
...